+++ gui/src/components/OnboardingCard/OnboardingQuickstartTab.tsx
@@ -3,8 +3,7 @@ import { Button } from "..";
 function OnboardingQuickstartTab() {
   return (
     <div className="flex justify-center items-center w-full h-full">
-      <div className="flex flex-col items-center justify-center w-1/2">
-        {" "}
+      <div className="flex flex-col items-center justify-center w-3/4">
         <h1 className="text-2xl font-bold text-center">Welcome to Continue</h1>
         <p className="leading-relaxed text-center">
           Let's find the setup that works best for you. You can always update
