+++ gui/src/pages/onboarding/ApiKeysOnboarding.tsx
@@ -5,7 +5,7 @@ import QuickModelSetup from "../../components/modelSelection/quickSetup/QuickMod
 import Toggle from "../../components/modelSelection/Toggle";
 import { IdeMessengerContext } from "../../context/IdeMessenger";
 import { getLocalStorage } from "../../util/localStorage";
-import DefaultModelConfigForm from "./DefaultModelConfigForm";
+import DefaultModelConfigForm from "../../components/OnboardingCard/DefaultModelConfigForm";
 import { useCompleteOnboarding } from "./utils";
 
 function ApiKeysOnboarding() {
